Implement PDF Appetizer version 2.0
Provide a PDF teaser to promote PDF usage.
Implement an addition for the logic of the PDF link:
Either link the PDF file (just as before) or - new -
present the visitor with a link to an information page
with further instructions on how to configure PDF creation
for any given manual.

Also add comments to code and remove some unused or broken stuff.